Introduce a Middleware trait.
The Middleware trait abstracts the composition of services and
middleware. Rather than each middleware component taking its upstream
service in its own manner, all of them receive a reference to an
upstream service as a part of their definition of call.

Some notes about this design:

The Middleware is parametric over Service, which allows users to
specify which sorts of Service their Middleware can wrap (for example,
if they need to implement another trait, or need to have a particular
Request, Response, or Error type).

The Middleware receives its upstream Service as an &Arc<S>. The inner
Service is stored in an Arc so that it can be cloned and returned as
part of the state of the future returned by this middleware (e.g. by
moving it into an AndThen call). However, it is passed by reference,
so that reference count need not be incremented.

For convenience, Middleware, Service, and their factory traits all have
a pre-implemented wrap method, which takes this middleware or service
and its corresponding pair, and construct a new service which wraps
the middleware around the service.  The return value of wrap implements
Service or NewService appropriately, so that it can continue to be
subjected to wrapping actions.

This allows users to conveniently write middleware chains, as in:

```rust
some_service.wrap(middleware1).wrap(middleware2).wrap(middleware3)
```